requirements.txt
Flask==2.1.2
numpy==1.23.5
opencv-python==4.7.0.72
matplotlib==3.7.1
seaborn==0.12.2
tensorflow==2.12.0
Pillow==9.5.0
